+++ Editor/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Unity.Meta.InstantGames.Sdk.Editor.Tests")]