+++ Runtime/FacebookInstantGamesSDK/Runtime/JsNative/JsRefArray.cs
@@ -55,6 +55,6 @@ namespace JsNative
             return values;
         }
 
-        protected override string CreateJs() => CreateArray();
+        internal override string CreateJs() => CreateArray();
     }
 }